scraping/pyproject.toml
[project]
name = "scholarhub-scraping"
version = "0.1.0"
description = "ScholarHub scholarship scraping pipeline"
requires-python = ">=3.12"
dependencies = [
    "convex>=0.7.0",
    "python-dotenv>=1.2.2",
    "scrapy>=2.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scholarhub_scraping"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "D",
    "ANN",
    "S",
    "B",
    "A",
    "C4",
    "T20",
    "RUF",
]
ignore = [
    "D100",
    "D104",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "D103", "ANN"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
